% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{as_tibble.po_datasets}
\alias{as_tibble.po_datasets}
\title{Convert po_datasets back to regular tibble}
\usage{
as_tibble.po_datasets(x, ...)
}
\description{
Convert po_datasets back to regular tibble
}
