% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{as_tibble.po_resources}
\alias{as_tibble.po_resources}
\title{Convert po_resources back to regular tibble}
\usage{
as_tibble.po_resources(x, ...)
}
\description{
Convert po_resources back to regular tibble
}
