% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear package cache}
\usage{
clear_cache(type = c("all", "api", "catalog", "files"))
}
\arguments{
\item{type}{Character. Type of cache to clear: "all" (default), "api", "catalog", or "files"}
}
\value{
Invisible TRUE
}
\description{
Clear package cache
}
