% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{knit_print.po_resources}
\alias{knit_print.po_resources}
\title{Knit print method for po_resources (for Quarto/markdown)}
\usage{
\method{knit_print}{po_resources}(x, max_items = getOption("peruopen.print_max", 10), ...)
}
\arguments{
\item{x}{A po_resources object}

\item{max_items}{Maximum number of items to display}

\item{...}{Additional arguments passed to knit_print methods}
}
\description{
Knit print method for po_resources (for Quarto/markdown)
}
