% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{knit_print.po_search_result}
\alias{knit_print.po_search_result}
\title{Knit print method for po_search results (for Quarto/markdown)}
\usage{
\method{knit_print}{po_search_result}(x, ...)
}
\arguments{
\item{x}{A po_search_result object}

\item{...}{Additional arguments passed to knit_print methods}
}
\description{
Knit print method for po_search results (for Quarto/markdown)
}
