% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peruopen.R
\docType{package}
\name{peruopen-package}
\alias{peruopen}
\alias{peruopen-package}
\title{peruopen: Access Peru's Open Data Portal (CKAN)}
\description{
This package provides a comprehensive interface to Peru's CKAN data portal
(datosabiertos.gob.pe), enabling users to query and search datasets,
extract metadata, and download resources directly into R.
}
\section{Main functions}{

\describe{
\item{\code{\link{po_catalog}}}{Get complete data catalog}
\item{\code{\link{po_search}}}{Search datasets and resources}
\item{\code{\link{po_get}}}{Smart getter for datasets/resources}
\item{\code{\link{po_explore}}}{Interactive data exploration}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/horaciochacon/peruopen}
  \item \url{https://horaciochacon.github.io/peruopen/}
  \item Report bugs at \url{https://github.com/horaciochacon/peruopen/issues}
}

}
\author{
\strong{Maintainer}: Horacio Chacón-Torrico \email{hchacont@uw.edu}

}
\keyword{internal}
