% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_datasets.R
\name{po_get_dataset_count}
\alias{po_get_dataset_count}
\title{Get total number of datasets}
\usage{
po_get_dataset_count(use_cache = TRUE)
}
\arguments{
\item{use_cache}{Logical. Whether to use cached results (default TRUE)}
}
\value{
Integer count of total datasets
}
\description{
Get total number of datasets
}
\keyword{internal}
