% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{po_get_resource_metadata}
\alias{po_get_resource_metadata}
\title{Get metadata for a specific resource}
\usage{
po_get_resource_metadata(resource_id)
}
\arguments{
\item{resource_id}{Character. The resource ID}
}
\value{
A list with resource metadata
}
\description{
Get metadata for a specific resource
}
\keyword{internal}
