% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{po_get_resource_url}
\alias{po_get_resource_url}
\title{Get the download URL for a resource}
\usage{
po_get_resource_url(resource_id)
}
\arguments{
\item{resource_id}{Character. The resource ID}
}
\value{
Character URL for the resource
}
\description{
Get the download URL for a resource
}
