% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_datasets.R
\name{po_list_datasets}
\alias{po_list_datasets}
\title{List all available datasets from Peru's CKAN portal}
\usage{
po_list_datasets(use_cache = TRUE)
}
\arguments{
\item{use_cache}{Logical. Whether to use cached results (default TRUE)}
}
\value{
A tibble with dataset names
}
\description{
List all available datasets from Peru's CKAN portal
}
