% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_datasets.R
\name{po_list_organizations}
\alias{po_list_organizations}
\title{Get available organizations/groups from CKAN portal}
\usage{
po_list_organizations(use_cache = TRUE)
}
\arguments{
\item{use_cache}{Logical. Whether to use cached results (default TRUE)}
}
\value{
A tibble with organization information
}
\description{
Get available organizations/groups from CKAN portal
}
