% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{print.po_datasets}
\alias{print.po_datasets}
\title{Print method for datasets tibbles}
\usage{
\method{print}{po_datasets}(x, max_items = getOption("peruopen.print_max", 10), ...)
}
\description{
Print method for datasets tibbles
}
