% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{as_tibble.po_datasets}
\alias{as_tibble.po_datasets}
\title{Convert po_datasets back to regular tibble}
\usage{
\method{as_tibble}{po_datasets}(x, ...)
}
\arguments{
\item{x}{A po_datasets object}

\item{...}{Additional arguments passed to as_tibble}
}
\description{
Convert po_datasets back to regular tibble
}
