% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{as_tibble.po_resources}
\alias{as_tibble.po_resources}
\title{Convert po_resources back to regular tibble}
\usage{
\method{as_tibble}{po_resources}(x, ...)
}
\arguments{
\item{x}{A po_resources object}

\item{...}{Additional arguments passed to as_tibble}
}
\description{
Convert po_resources back to regular tibble
}
