% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{print.po_explore_result}
\alias{print.po_explore_result}
\title{Print method for po_explore results}
\usage{
\method{print}{po_explore_result}(x, ...)
}
\arguments{
\item{x}{A po_explore_result object}

\item{...}{Additional arguments passed to print methods}
}
\description{
Print method for po_explore results
}
