% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_api.R
\name{print.po_resources}
\alias{print.po_resources}
\title{Print method for resources tibbles}
\usage{
\method{print}{po_resources}(x, max_items = getOption("peruopen.print_max", 10), ...)
}
\arguments{
\item{x}{A po_resources object}

\item{max_items}{Maximum number of items to display}

\item{...}{Additional arguments passed to print methods}
}
\description{
Print method for resources tibbles
}
